+++ java-security/src/main/java/com/sap/cloud/security/token/validation/validators/JsonWebKeyImpl.java
@@ -8,37 +8,23 @@ package com.sap.cloud.security.token.validation.validators;
 import com.sap.cloud.security.xsuaa.Assertions;
 
 import jakarta.annotation.Nullable;
-import java.math.BigInteger;
-import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
 import java.security.PublicKey;
 import java.security.spec.InvalidKeySpecException;
-import java.security.spec.KeySpec;
-import java.security.spec.RSAPublicKeySpec;
-import java.security.spec.X509EncodedKeySpec;
-import java.util.Base64;
 import java.util.Objects;
-
-import static com.sap.cloud.security.token.validation.validators.JsonWebKeyConstants.BEGIN_PUBLIC_KEY;
-import static com.sap.cloud.security.token.validation.validators.JsonWebKeyConstants.END_PUBLIC_KEY;
 
 class JsonWebKeyImpl implements JsonWebKey {
 	private final JwtSignatureAlgorithm keyAlgorithm;
 	private final String keyId;
-	private final String pemEncodedPublicKey;
-	private final String modulus;
-	private final String publicExponent;
+	private final KeyMaterial keyMaterial;
 	private PublicKey publicKey;
 
-	JsonWebKeyImpl(JwtSignatureAlgorithm keyAlgorithm, @Nullable String keyId, String modulus,
-			String publicExponent, @Nullable String pemEncodedPublicKey) {
+	JsonWebKeyImpl(JwtSignatureAlgorithm keyAlgorithm, @Nullable String keyId, KeyMaterial keyMaterial) {
 		Assertions.assertNotNull(keyAlgorithm, "keyAlgorithm must be not null");
-		this.keyId = keyId != null ? keyId : DEFAULT_KEY_ID;
-
-		this.pemEncodedPublicKey = pemEncodedPublicKey;
-		this.publicExponent = publicExponent;
-		this.modulus = modulus;
+		Assertions.assertNotNull(keyMaterial, "keyMaterial must be not null");
 		this.keyAlgorithm = keyAlgorithm;
+		this.keyId = keyId != null ? keyId : DEFAULT_KEY_ID;
+		this.keyMaterial = keyMaterial;
 	}
 
 	@Override
@@ -54,49 +40,15 @@ class JsonWebKeyImpl implements JsonWebKey {
 
 	@Override
 	public PublicKey getPublicKey() throws NoSuchAlgorithmException, InvalidKeySpecException {
-		if (publicKey != null) {
-			return publicKey;
-		}
-		if (pemEncodedPublicKey != null) {
-			publicKey = createPublicKeyFromPemEncodedPublicKey(keyAlgorithm, pemEncodedPublicKey);
-		} else if (keyAlgorithm.type().equalsIgnoreCase("RSA")) {
-			publicKey = createRSAPublicKey(publicExponent, modulus);
-		} else {
-			throw new IllegalStateException("JWT token with web key type " + keyAlgorithm + " can not be verified.");
+		if (publicKey == null) {
+			publicKey = keyMaterial.toPublicKey(keyAlgorithm);
 		}
 		return publicKey;
 	}
 
-	static PublicKey createRSAPublicKey(String publicExponent, String modulus)
-			throws NoSuchAlgorithmException, InvalidKeySpecException {
-		BigInteger n = new BigInteger(1, Base64.getUrlDecoder().decode(modulus));
-		BigInteger e = new BigInteger(1, Base64.getUrlDecoder().decode(publicExponent));
-		KeySpec keySpec = new RSAPublicKeySpec(n, e);
-
-		KeyFactory keyFactory = KeyFactory.getInstance("RSA");
-		return keyFactory.generatePublic(keySpec);
-	}
-
-	public static PublicKey createPublicKeyFromPemEncodedPublicKey(JwtSignatureAlgorithm algorithm,
-			String pemEncodedKey)
-			throws NoSuchAlgorithmException, InvalidKeySpecException {
-		byte[] decodedBytes = Base64.getMimeDecoder().decode(convertPEMKey(pemEncodedKey));
-
-		X509EncodedKeySpec keySpecX509 = new X509EncodedKeySpec(decodedBytes);
-		KeyFactory keyFactory = KeyFactory.getInstance(algorithm.type());
-		return keyFactory.generatePublic(keySpecX509);
-	}
-
-	public static String convertPEMKey(String pemEncodedKey) {
-		String key = pemEncodedKey;
-		key = key.replace(BEGIN_PUBLIC_KEY, "");
-		key = key.replace(END_PUBLIC_KEY, "");
-		return key;
-	}
-
 	@Override
 	public int hashCode() {
-		return Objects.hash(keyAlgorithm, keyId != null ? keyId : DEFAULT_KEY_ID);
+		return Objects.hash(keyAlgorithm, keyId);
 	}
 
 	@Override